src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/gesture-tap-remove-node-crash-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
PASS unless crash.
